# Fuse inner loop kernels in device CKF