… device/common/CMakeLists.txt
… device/common/include/traccc/finding/device/add_links_for_holes.hpp
… device/common/include/traccc/finding/device/count_measurements.hpp
… device/common/include/traccc/finding/device/impl/add_links_for_holes.ipp
… device/common/include/traccc/finding/device/impl/count_measurements.ipp
… device/cuda/src/finding/finding_algorithm.cu