@stephenswat 2024-09-20
PR updated.

> Have you checked the performance of this?

It's essentially identical:

**Before**:

```
Reconstructed track parameters: 3061231
Time totals:
                  File reading  80234 ms
            Warm-up processing  984 ms
              Event processing  8227 ms
Throughput:
            Warm-up processing  98.4611 ms/event, 10.1563 events/s
              Event processing  82.2799 ms/event, 12.1536 events/s
```

**After:**

```
Reconstructed track parameters: 3063909
Time totals:
                  File reading  79725 ms
            Warm-up processing  27888 ms
              Event processing  8229 ms
Throughput:
            Warm-up processing  2788.83 ms/event, 0.358573 events/s
              Event processing  82.2956 ms/event, 12.1513 events/s
```